Improve Release page user experience.
The first draft of the Release page was, to put it mildly, bad.  This
commit introduces a little redesign / reorganization to the workflows
which should improve the whole experience.  Releases now refresh at
certain cues, and will clear their working cache as well, within an app
and component context.
