+++ lib/notifications/NotificationBar.component.js
@@ -11,7 +11,7 @@ const Notification = ({ notification, onClick }) =>
         { notification.get('message') }
       </div>
       <div className='notification-bar-close'>
-        <a href='#' onClick={e => { onClick() }} className='glyph glyph-x' />
+        <a href='#' onClick={e => { onClick(e) }} className='glyph glyph-x' />
       </div>
     </div>
   </div>
