+++ lib/releases/ReleaseMount.component.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ReleaseMount = ({ mount }) =>
+  <tr>
+    <td>{ mount.get('volume') }</td>
+    <td>{ mount.get('path') }</td>
+  </tr>
+
+ReleaseMount.propTypes = { mount: React.PropTypes.object.isRequired }
+
+export default ReleaseMount
