+++ lib/releases/ReleaseVariable.component.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ReleaseVariable = ({ variable }) =>
+  <tr>
+    <td>{ variable.get('name') }</td>
+    <td>{ variable.get('value') }</td>
+  </tr>
+
+ReleaseVariable.propTypes = { variable: React.PropTypes.object.isRequired }
+
+export default ReleaseVariable
