+++ lib/releases/Releases.component.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import Release from './Release.component'
+import FadeChange from '../shared/FadeChange.animation'
 import NotFound from '../shared/NotFound.component'
 
 export default class Releases extends React.Component {
@@ -37,29 +38,34 @@ export default class Releases extends React.Component {
             Releases
           </div>
           <div className='context-menu'>
+            <FadeChange>
+              {
+                target && target.toList().count() > 0 && (
+                  <button onClick={ deploy }
+                          className='with-glyph glyph-deploy-action-color'>
+                    Deploy Changes
+                  </button>
+                )
+              }
 
-            {
-              target && target.toList().count() > 0 && (
-                <button onClick={ deploy }
-                        className='with-glyph glyph-deploy-action-color'>
-                  Deploy Changes
-                </button>
-              )
-            }
+              {
+                target && target.toList().count() === 0 && (
+                  <button className='with-glyph glyph-plus-action-color'
+                          onClick={ create }>
+                    Create a Release
+                  </button>
+                )
+              }
 
-            {
-              target && target.toList().count() === 0 && (
-                <button className='with-glyph glyph-plus-action-color'
-                        onClick={ create }>
-                  Create a Release
-                </button>
-              )
-            }
-
-            <button className='with-glyph glyph-x-action-color'
-                    onClick={ destroyTarget }>
-              Clear Changes
-            </button>
+              {
+                target && target.toList().count() > 0 && (
+                  <button className='with-glyph glyph-x-action-color'
+                          onClick={ destroyTarget }>
+                    Clear Changes
+                  </button>
+                )
+              }
+            </FadeChange>
           </div>
         </header>
