+++ lib/releases/releases.actions.js
@@ -46,6 +46,7 @@ export const destroy = (id, appName, componentName) => (
 
 export const Remove = "releases:remove"
 export const Insert = "releases:insert"
+export const Clear = "releases:clear"
 
 export const remove = (id, appName, componentName) => (
   { type: Remove, id, appName, componentName }
@@ -53,4 +54,8 @@ export const remove = (id, appName, componentName) => (
 
 export const insert = (id, release, appName, componentName) => (
   { type: Insert, id, release, appName, componentName }
+)
+
+export const clear = (appName, componentName) => (
+  { type: Clear, appName, componentName }
 )
