+++ lib/releases/releases.reducers.js
@@ -9,6 +9,8 @@ export default function releases(state = fromJS({}), action) {
   ].join('-')
 
   switch (action.type) {
+    case ReleaseActions.Clear:
+      return fromJS({})
     case ReleaseActions.Insert:
       return state.set(releaseKey, action.release)
     case ReleaseActions.Remove:
