+++ lib/releases/releases.sagas.js
@@ -124,8 +124,7 @@ export function* createRelease(getState, { params, appName, componentName }) {
     let id = body.timestamp
 
     yield fork(infoMessages, `Created release '${id}'`)
-    yield getRelease({ id, appName, componentName })
-    yield put(push(`/apps/${appName}/components/${componentName}/releases`))
+    yield fork(refreshReleases, { appName, componentName })
   } else {
     if (body.error) {
       yield fork(errorMessages, body.error)
@@ -133,6 +132,13 @@ export function* createRelease(getState, { params, appName, componentName }) {
       yield fork(errorMessages, body)
     }
   }
+}
+
+function* refreshReleases({ appName, componentName }) {
+  yield call(() => new Promise(resolve => setTimeout(resolve, 300)))
+  yield put(ReleaseActions.fetch(appName, componentName))
+  yield put(ReleaseActions.getCurrent(appName, componentName))
+  yield put(ReleaseActions.getTarget(appName, componentName))
 }
 
 function* deleteRelease({ id, appName, componentName }) {
@@ -145,6 +151,7 @@ function* deleteRelease({ id, appName, componentName }) {
 
   if (response.status === 200) {
     yield put(ReleaseActions.remove(id, appName, componentName))
+    yield put(ReleaseActions.clear(appName, componentName))
     yield fork(infoMessages, `Deleted release ${id}`)
   } else {
     if (body.error) {
@@ -179,7 +186,12 @@ function* destroy(getState) {
   yield takeEvery(ReleaseActions.Destroy, deleteRelease)
 }
 
+function* clear(getState) {
+  yield takeEvery(ReleaseActions.Clear, refreshReleases)
+}
+
 export function* all(getState) {
+  yield fork(clear, getState)
   yield fork(fetch, getState)
   yield fork(get, getState)
   yield fork(create, getState)
